+++ core/src/libtx/error.rs
@@ -16,7 +16,7 @@
 use failure::{Backtrace, Context, Fail};
 use std::fmt::{self, Display};
 
-use crate::core::{committed, transaction};
+use crate::core::transaction;
 use crate::keychain;
 use crate::util::secp;
 
@@ -44,12 +44,6 @@ pub enum ErrorKind {
 	/// Rangeproof error
 	#[fail(display = "Rangeproof Error")]
 	RangeProof(String),
-	/// Fee error
-	#[fail(display = "Fee Error")]
-	Fee(String),
-	/// Error from summing commitments via committed trait.
-	#[fail(display = "Committed Error")]
-	Committed(committed::Error),
 }
 
 impl Fail for Error {
@@ -93,14 +87,6 @@ impl From<secp::Error> for Error {
 	fn from(error: secp::Error) -> Error {
 		Error {
 			inner: Context::new(ErrorKind::Secp(error)),
-		}
-	}
-}
-
-impl From<committed::Error> for Error {
-	fn from(error: committed::Error) -> Error {
-		Error {
-			inner: Context::new(ErrorKind::Committed(error)),
 		}
 	}
 }
