+++ core/src/libtx/mod.rs
@@ -27,7 +27,6 @@ mod error;
 pub mod proof;
 pub mod reward;
 pub mod secp_ser;
-pub mod slate;
 
 use crate::consensus;
 use crate::core::Transaction;
