+++ wallet/src/adapters/http.rs
@@ -14,7 +14,7 @@
 
 use crate::api;
 use crate::controller;
-use crate::core::libtx::slate::Slate;
+use crate::libwallet::slate::Slate;
 use crate::libwallet::{Error, ErrorKind};
 use crate::{instantiate_wallet, HTTPNodeClient, WalletCommAdapter, WalletConfig};
 /// HTTP Wallet 'plugin' implementation
