+++ wallet/src/adapters/mod.rs
@@ -22,7 +22,7 @@ pub use self::http::HTTPWalletCommAdapter;
 pub use self::keybase::KeybaseWalletCommAdapter;
 pub use self::null::NullWalletCommAdapter;
 
-use crate::core::libtx::slate::Slate;
+use crate::libwallet::slate::Slate;
 use crate::libwallet::Error;
 use crate::WalletConfig;
 use std::collections::HashMap;
