+++ wallet/src/adapters/null.rs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 /// Null Output 'plugin' implementation
-use crate::core::libtx::slate::Slate;
+use crate::libwallet::slate::Slate;
 use crate::libwallet::Error;
 use crate::{WalletCommAdapter, WalletConfig};
