+++ wallet/src/libwallet/internal/selection.rs
@@ -15,10 +15,11 @@
 //! Selection of inputs for building transactions
 
 use crate::core::core::{amount_to_hr_string, Transaction};
-use crate::core::libtx::{build, slate::Slate, tx_fee};
+use crate::core::libtx::{build, tx_fee};
 use crate::keychain::{Identifier, Keychain};
 use crate::libwallet::error::{Error, ErrorKind};
 use crate::libwallet::internal::keys;
+use crate::libwallet::slate::Slate;
 use crate::libwallet::types::*;
 use std::collections::HashMap;
 use std::marker::PhantomData;
