+++ wallet/src/libwallet/types.rs
@@ -21,11 +21,11 @@ use crate::core::libtx::aggsig;
 use crate::core::ser;
 use crate::keychain::{Identifier, Keychain};
 use crate::libwallet::error::{Error, ErrorKind};
+use crate::libwallet::slate::ParticipantMessages;
 use crate::util::secp::key::{PublicKey, SecretKey};
 use crate::util::secp::{self, pedersen, Secp256k1};
 use chrono::prelude::*;
 use failure::ResultExt;
-use grin_core::libtx::slate::ParticipantMessages;
 use serde;
 use serde_json;
 use std::collections::HashMap;
