+++ wallet/tests/repost.rs
@@ -17,8 +17,8 @@ extern crate log;
 
 use self::core::global;
 use self::core::global::ChainTypes;
-use self::core::libtx::slate::Slate;
 use self::keychain::ExtKeychain;
+use self::libwallet::slate::Slate;
 use self::wallet::test_framework::{self, LocalWalletClient, WalletProxy};
 use self::wallet::{libwallet, FileWalletCommAdapter};
 use grin_core as core;
